+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2026-03
+
+* Context menu to edit line vertexes
+
 ## 2026-02
 
 * Support Wikipedia layer